REFACTOR: Remove bitdubai from wallet assets issuer wallet plugin. #5807
