+++ DAP/plugin/wallet/fermat-dap-plugin-wallet-assets-issuer-wallet/build.gradle
